backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.2
bson==0.5.10
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
